+++ NEWS.md
@@ -43,6 +43,7 @@
 ## Minor changes
 
 * Better error and verbose messages (#422, #423).
+* The default mouse mode for `wheel` has been changed to `"pull"`.
 
 # eplusr 0.14.1